requirements.txt
python-telegram-bot[job-queue]
python-dotenv
pypdf
httpx
openai
